cygwin: petty changes to ascertain build under cygwin

+++ chipmunk/util.c
@@ -667,7 +667,7 @@ a2time( const char* str, time_t* t, time_t from )
             continue;
         }
 
-        if( !isdigit( str[n] ) ) return ERR_NONDIGIT;
+        if( !isdigit( (int)str[n] ) ) return ERR_NONDIGIT;
 
         if( new_field ) {
             new_field = 0;
@@ -763,7 +763,7 @@ a2double( const char* str, double* pval )
     assert( str );
 
     /* skip to the first */
-    for( p = str; *p && !isalpha(*p); ++p );
+    for( p = str; *p && !isalpha((int)*p); ++p );
 
     if( '\0' != *p ) {
         /* there is a modifier, calculate multiplication
